feat(worklist): a group can be opened, and the classifier splits at its seam
The batch row draws as what it is — "43 likely automated senders" — names a
few of its members so a reader can check it before answering it, and offers
the way in. A group nobody can see into is a group nobody trusts, and an
untrusted group is worse than the pile it replaced.

classify.go passed the file-length ceiling, and the seam was the right one:
everything moved answers a question the rest do not — whether a decision
holds up a person, or is the routine tidying that groups.
